requirements.txt
requests
pandas
flask
modules
psycopg2-binary